# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Telegram Notify Bot API",
    "description": "API criada utilizando a lib npm `node-telegram-bot-api`(https://www.npmjs.com/package/node-telegram-bot-api). Com o objetivo de disponibilizar os recursos por meio de API REST.",
    "version": "1.0.0"
  },
  "host": "telegram-notify-api.achimid.com.br",
  "basePath": "/api/v1",
  "schemes": [
    "https"
  ],
  "paths": {
    "/bot": {
      "post": {
        "tags": [
          "Bot"
        ],
        "summary": "Cadastro de um Bot",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "msgSuccess": {
                  "example": "any"
                },
                "msgDuplicated": {
                  "example": "any"
                },
                "msgError": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/bot/{token}": {
      "put": {
        "tags": [
          "Bot"
        ],
        "summary": "Alterar as informações do Bot",
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "msgSuccess": {
                  "example": "any"
                },
                "msgDuplicated": {
                  "example": "any"
                },
                "msgError": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "Bot"
        ],
        "summary": "Excluir um Bot já cadastrado",
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "get": {
        "tags": [
          "Bot"
        ],
        "summary": "Recuperar as informações de um Bot",
        "description": "",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/message/send": {
      "post": {
        "tags": [
          "Message"
        ],
        "summary": "Efetuar o envio de uma mensagem para um usuário[chat.id] específico",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                },
                "id": {
                  "example": "any"
                },
                "text": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/message/send/all": {
      "post": {
        "tags": [
          "Message"
        ],
        "summary": "Efetuar o envio de uma mensagem broadcast para todos os usuários[chat.id] de um Bot",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "example": "any"
                },
                "text": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  }
}